ref: move digests Digest / Notification / Record types
